Add a test for postcodes

+++ postcodes.rb
@@ -8,8 +8,21 @@ require 'csv'
 require 'mysql'
 require 'configuration'
 require 'people'
+require 'optparse'
 
-conf = Configuration.new
+options = {}
+
+OptionParser.new do |opts|
+  opts.banner = "Usage: postcodes.rb [--test]"
+
+  opts.on("--test", "Run in test mode (no DB updates)") do |test|
+    options[:test] = test
+  end
+end.parse!
+
+unless options[:test]
+  conf = Configuration.new
+end
 
 def quote_string(s)
   s.gsub(/\\/, '\&\&').gsub(/'/, "''") # ' (for ruby-mode)
@@ -29,10 +42,14 @@ constituencies.each do |constituency|
   throw "Constituency #{constituency} not found" unless all_members.any? {|m| m.division == constituency}
 end
 
-db = Mysql.real_connect(conf.database_host, conf.database_user, conf.database_password, conf.database_name)
+if options[:test]
+  puts "Postcodes look good!"
+else
+  db = Mysql.real_connect(conf.database_host, conf.database_user, conf.database_password, conf.database_name)
 
-# Clear out the old data
-db.query("DELETE FROM postcode_lookup")
+  # Clear out the old data
+  db.query("DELETE FROM postcode_lookup")
 
-values = data.map {|row| "('#{row[0]}', '#{quote_string(row[1])}')" }.join(',')
-db.query("INSERT INTO postcode_lookup (postcode, name) VALUES #{values}")
+  values = data.map {|row| "('#{row[0]}', '#{quote_string(row[1])}')" }.join(',')
+  db.query("INSERT INTO postcode_lookup (postcode, name) VALUES #{values}")
+end
